#41111 Switch SendNotifyEmail to use Hangfire jobs rather than an Azure queue
